authOption lib file create
